Crate forecast

This module exposes an SDK for interacting with the Dark Sky API.

Structs

Alert
Model object representing a severe weather warning issued by a government authority for the requested location.

ApiClient
The ApiClient is a thin wrapper around a hyper::Client which sends requests to the Forecast and Time Machine APIs.

ApiResponse
Model object representing a Forecast or Time Machine API response.

DataBlock
Model object representing the various weather phenomena ocurring over a period of time.

DataPoint
Model object containing various properties, each represenging the average (unless otherwise specified) of a particular weather phenomenon ocurring during a period of time.

Flags
Model object representing a flag which contains miscellaneous metadata about a request.

ForecastRequest
Model object representing a request to the Forecast API.

ForecastRequestBuilder
Builder object used to construct a ForecastRequest.

TimeMachineRequest
Model object representing a request to the Time Machine API.

TimeMachineRequestBuilder
Builder object used to construct a TimeMachineRequest.

Enums

ExcludeBlock
Model object representing a DataBlock to exclude from the response.

ExtendBy
When present in a request, this feature causes response data to be reported for 168 hours into the future instead of 48 hours.

Icon
Model object representing an icon for display.

Lang
Model object representing language.

PrecipType
Model object representing the kind of precipitation ocurring at a particular time.

Units
Model object representing measurement units.