Struct forecast::ForecastRequestBuilder

pub struct ForecastRequestBuilder<'a> { /* fields omitted */ }

Builder object used to construct a ForecastRequest.

Methods

impl<'a> ForecastRequestBuilder<'a>

fn new(api_key: &'a str, latitude: f64, longitude: f64) -> ForecastRequestBuilder

A Forecast API request is constructed with required params api_key, latitude, and longitude.

fn exclude_block(self, exclude_block: ExcludeBlock) -> ForecastRequestBuilder<'a>

Add a DataBlock to exclude from the response.

fn exclude_blocks(self, exclude_blocks: &mut Vec<ExcludeBlock>) -> ForecastRequestBuilder<'a>

Add multiple DataBlocks to exclude from the response.

fn extend(self, extend: ExtendBy) -> ForecastRequestBuilder<'a>

Extend the time window of the response data from 48 hours to 168 hours.

fn lang(self, lang: Lang) -> ForecastRequestBuilder<'a>

Set the language for messages in the response data.

fn units(self, units: Units) -> ForecastRequestBuilder<'a>

Set the measurement units for response data.

fn build(self) -> ForecastRequest<'a>

Finalize the request.