Struct forecast::DataPoint

pub struct DataPoint {
    pub apparent_temperature: Option<f64>,
    pub apparent_temperature_high: Option<f64>,
    pub apparent_temperature_high_time: Option<u64>,
    pub apparent_temperature_low: Option<f64>,
    pub apparent_temperature_low_time: Option<u64>,
    pub apparent_temperature_max: Option<f64>,
    pub apparent_temperature_max_time: Option<u64>,
    pub apparent_temperature_min: Option<f64>,
    pub apparent_temperature_min_time: Option<u64>,
    pub cloud_cover: Option<f64>,
    pub dew_point: Option<f64>,
    pub humidity: Option<f64>,
    pub icon: Option<Icon>,
    pub moon_phase: Option<f64>,
    pub nearest_storm_bearing: Option<f64>,
    pub nearest_storm_distance: Option<f64>,
    pub ozone: Option<f64>,
    pub precip_accumulation: Option<f64>,
    pub precip_intensity: Option<f64>,
    pub precip_intensity_max: Option<f64>,
    pub precip_intensity_max_time: Option<u64>,
    pub precip_probability: Option<f64>,
    pub precip_type: Option<PrecipType>,
    pub pressure: Option<f64>,
    pub summary: Option<String>,
    pub sunrise_time: Option<u64>,
    pub sunset_time: Option<u64>,
    pub temperature: Option<f64>,
    pub temperature_high: Option<f64>,
    pub temperature_high_time: Option<u64>,
    pub temperature_low: Option<f64>,
    pub temperature_low_time: Option<u64>,
    pub temperature_max: Option<f64>,
    pub temperature_max_time: Option<u64>,
    pub temperature_min: Option<f64>,
    pub temperature_min_time: Option<u64>,
    pub time: u64,
    pub uv_index: Option<f64>,
    pub uv_index_time: Option<u64>,
    pub visibility: Option<f64>,
    pub wind_bearing: Option<f64>,
    pub wind_gust: Option<f64>,
    pub wind_gust_time: Option<u64>,
    pub wind_speed: Option<f64>,
}

Model object containing various properties, each representing the average (unless otherwise specified) of a particular weather phenomenon occurring during a period of time.

Fields
apparent_temperature: Option<f64> apparent_temperature_high: Option<f64> apparent_temperature_high_time: Option<u64> apparent_temperature_low: Option<f64> apparent_temperature_low_time: Option<u64> apparent_temperature_max: Option<f64>

Deprecated since 1.0.0

apparent_temperature_max_time: Option<u64>

Deprecated since 1.0.0

apparent_temperature_min: Option<f64>

Deprecated since 1.0.0

apparent_temperature_min_time: Option<u64>

Deprecated since 1.0.0

cloud_cover: Option<f64> dew_point: Option<f64> humidity: Option<f64> icon: Option<Icon> moon_phase: Option<f64> nearest_storm_bearing: Option<f64> nearest_storm_distance: Option<f64> ozone: Option<f64> precip_accumulation: Option<f64> precip_intensity: Option<f64> precip_intensity_max: Option<f64> precip_intensity_max_time: Option<u64> precip_probability: Option<f64> precip_type: Option<PrecipType> pressure: Option<f64> summary: Option<String> sunrise_time: Option<u64> sunset_time: Option<u64> temperature: Option<f64> temperature_high: Option<f64> temperature_high_time: Option<u64> temperature_low: Option<f64> temperature_low_time: Option<u64> temperature_max: Option<f64>

Deprecated since 1.0.0

temperature_max_time: Option<u64>

Deprecated since 1.0.0

temperature_min: Option<f64>

Deprecated since 1.0.0

temperature_min_time: Option<u64>

Deprecated since 1.0.0

time: u64 uv_index: Option<f64> uv_index_time: Option<u64> visibility: Option<f64> wind_bearing: Option<f64> wind_gust: Option<f64> wind_gust_time: Option<u64> wind_speed: Option<f64>